keyword_sync/requirements.txt
gspread==5.7.2
google-auth==2.20.0
redis>=4.5.0
python-dotenv==1.0.0
